+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed diagnostics not showing when working in a new file with no workspace open
 - Fixed race condition where sometimes the server does not receive user configuration on initial load, falling back to defaults
 - `luau-lsp.fflags.override` will now be parsed when using CLI analyze settings. Note that the other `fflags` options are not supported in CLI analyze mode
+- Fixed semantic tokens segfault crash on some tables
 
 ## [1.22.0] - 2023-06-30